files/backend/requirements.txt
# ─── Web ─────────────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0

# ─── Database ─────────────────────────────────────────────────────────────────
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1

# ─── Validation / Settings ───────────────────────────────────────────────────
pydantic==2.7.1
pydantic-settings==2.2.1
email-validator==2.1.1

# ─── Auth ─────────────────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ─── Cache / Tasks ───────────────────────────────────────────────────────────
redis[hiredis]==5.0.4
celery==5.4.0

# ─── Storage ─────────────────────────────────────────────────────────────────
boto3==1.34.110
Pillow==10.3.0

# ─── Email ───────────────────────────────────────────────────────────────────
resend==0.8.0

# ─── HTTP (internal calls / webhooks) ────────────────────────────────────────
httpx==0.27.0

# ─── Dev / Linting / Testing ─────────────────────────────────────────────────
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0
ruff==0.4.4
mypy==1.10.0
python-multipart==0.0.9
